Issue-478: duplicate publishPlugin information to android
The android component only builds its components in the afterEvaluate phase. Hence, the publishing configuration needs to be done in the afterEvaluate phase of that plugin and nowhere else. See: https://developer.android.com/studio/build/maven-publish-plugin

Because I saw that there was a new one and thought why not?
Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>