Issue-478: separate maven from gradle plugin publication
When using the catchall 'publish' tasks, gradle would execute the publication to maven and the publication to gradle in the same run. Since these two task produce a publication with the exact same coordinates, but different metadata (pom vs no pom) the latter would overwrite the former leading to a failure of the sonatype pom verification.

Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>