Issue-478: ensure the property finders output is handled correctly
It turns out that there is no way to return a string or an array with elements that themselves contain spaces in a way that bash will interpret the elements correctly (i.e. var1, var 2, var3 -> var1, var, 2, var3). The solution is to use bash's nameref feature. This allows a function to write to a variable directly without knowing the actual variable name.

Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>