Issue-558: group and schedule Renovate PRs
The massive amount of Renovate PRs generate a lot of noise. Out of the options to reduce that noise, I chose to first just group automerge PR and limit the time at which the will be opened. The most powerful noise-reduction feature, branch-automerge, is currently incompatible with the repos branch protection settings, also I am not yet sure for which types of updates I should activate that.

Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>