WIP create new subproject for spock2
spock2 has not been released yet and our package only uses a thin wrapper for junit4. Hence it is better to have a junit4 package supporting spock1 and a junit5 package that navtively supports junit5.

spock2 package is in development and is thus not scheduled for a release
also prevent renovatebot from attempting to update the packages in spock1

Build currently fails due to unrelated issue