replace every compile with implementation
Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>

remove temp dir from Junit 5 test

A file wrapper object had to be introduced to keep the structure of the test similar to the way it was before.

This is because JUnit5's @TempDir annotation only becomes active after the "postProcessTestInstance" phase, in which JGiven passes around the reference for the Scenario state. Using temp dir directly thus results in a null pointer exception

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>