make all tests work nicely with groovy 3.
The upgrade to gradle 7 makes gradle incompatible with spock-1.x, because that runs on groovy-2.5. Therefore an upgrade to spock-2 is necessary, which is based on Junit5.

 A main complication were the Temporary Folders, which have been replaced by a different mechanism in JUnit5. While simply reusing the Temporary Folder manually would have been possible, I

opted to use spock's tempdir annotation order to avoid potential suffering from future incompatabilities. Also it turned out that many of the "TemporaryFolder"'s capabilities can be easily achived with java's File object.

For the JUnit tests, a file wrapper object had to be introduced to keep the structure of the test similar to the way it was before.

This is because JUnit5's @TempDir annotation only becomes active after the "postProcessTestInstance" phase, in which JGiven passes around the reference for the Scenario state. Using temp dir directly thus results in a null pointer exception

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>