Issue-102: Reorganize class relations
Despite the stage state being factored out of the the executor. It remained closely linked, with the exector having write access to a field of the state. Such a relation is only approprate for an inner class. Thus, the responsibilities of the stage state were divided into the management of the lifecycle methods, which are quite independent from the executor and the management of the stages itsself, which are not.
Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>