Issue-648: document the nested steps annotation
The nested steps feature was not advertised in the documentation. It has been added as a subsection of the step methods section, because it didn't seem complex enough to warrant its own section. It has been added as a subsection so that the content may be found more easily when scanning the document

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>