increase priority of method selector in testng framework test
TestNG replaced the list of method selectors in their method inclusion method by a set.
Also TestNG adds a default XMLMethodSelector with priority 10 to the method selector list. This default XMLMethodSelector is always evaluated first.

Contrary to the list addition method the Set addition method uses a comparator that is attached to the selection method. This comparator only compares the priorities of the given Keys to check object equality. The employed TreeSet, or the backing treeMap, respectively exchange the value if two keys compare to equal. However, in the case of TreeSet that value is a dummy, so in effect the addition of a new method to a set with wherein the an object with the same priority already exists gets silently ignored.

The issue has been reported under https://github.com/cbeust/testng/issues/2595

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>