replace deprecated methods in gradle file
 - replace deprecated use of enabled property for jacoco report tasks
 - replace http with https in URLs
 - replace deprecated JavaExec methods

despite throwing several warnings, gradle proposes the useage of `mainClass` in their documentation for JavaExec methods. Thus, noinspection comments have been added to highlight that this is in fact the proposed method of usage. Since two noinspection comments cannot be placed on the same line, one had to be made a file-wide noinspection comment.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>