Delete gradle properties files in example projects
These files are potential symlinks. Git symlinks are represented as standard textfiles on windows by either git or IntelliJ with the link target being the content of the text file. This is very hard to recognize for a user.

The files will be replaced by actual text files in the next commit

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>