update bash scripts to allow update of several files
also write tests to ensure that the update function fails hard if the property to be updated is not found. This is to prevent a divergence between properties files and scripts that act on them without notice.

additionally update the test harness to notice a lack of input and also report the result via a return value.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>