Issue-271: unit test the configuration
* Make value holding the properties an instance field
* Make constructor private

The above measures make the singleton class more aligned with the singleton pattern while also enable unit testing of the config class.

Due to the amount of classes in the impl package a private constructor is preferred over a package private one as the latter might be seen as an invitation to just instantiate the class, when this is clearly not desired.

Eager reading of the properties file has been deemed acceptable, because the `static` block within the class will ensure that the file is read early anyway.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>