Alter Memory consumption test
according to Issue-294 there was a memory leak at the beginning of using ByteBuddy. The current test checked that the memory consumption always stayed the same, using the global JVM memory to calculate usage.
However, varying values of consumption, though not in an ever increasing fashion, were observed in our CI pipeline, thus failing this test.

We therefore change the assertion to mirror its assumed original goal more closely by just requiring that there is no overall increase in consumption over the runtime of the test.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>