Issue-755: implement awful solution
The source of the incorrect timings is that the StepInterceptor and the ScenarioModelBuilder have different concepts about what a step is. Consequently, in the eyes of the scenario builder, the stepMethodFinished is called way more often than it should be. This is the source of incorrect timings in the presence of nested steps, intro and filler words.

The current solution tries to track the and accomodate for the difference in understandings when it comes to timings.

The limitation on timing is part of what makes this implementation so awful.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>