Increase growth cutoff for ProxyClassPerformanceTest
Previously the test failed for a scenario where the reported memory usage would be constant but increased for the last recorded item. In such a case the reported growth would not be 0, but looking at the record would reveal no continuous growth.

Therefore we now allow the last item in the record to be an increment by 1.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>