consolidate and organize scripts
There are now quite the amount of bash-scripts in this repository so sorting them into files that execute, files that provide functions and files that test seemed reasonable.

also add a git attributes file such that the bash scripts may be executed via wsl on Windows.

Note: The two update version functions for scala and gradle MUST NOT be consolidated! This is because these scripts are used on MacOS and Unix. Unix uses GNU grep and GNU sed, wherea MacOS uses BSD sed and BSD grep. the GNU and BSD versions differ in their interpretations of REGEX and are only interchangeable up to the POSIX specification, which, with regards to REGEX is unfortunately quite limited. Instead of confusing the reader with a hard to read, virtually obsolete regex standard that needs to satisfy two OS, I opted to just duplicate the code and avoid regex altogether.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>

Revert version replacement back to function duplication

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>