Issue-821: Fix logging errors during annotation reading
In commit fd978254a2494e8dde4bc194ea42251647c8fef2 a catch clause was removed that omitted errors from the annotation evaluator that were from reading annotations that simply don't have a value method, because they were never meant to be tags.
This results in JGiven printing lots of errors for every annation that could possbliy exist on a test class or method.

Therefore the catch clause was reintroduced an a test for its existence was added.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>