Fix global logging for test log handler
The test logger uses a non threadsafe static list to record log events that appear on any class to which it is attached.

A static field is not required for the purposes of this class, which is recording log events during a single run of a single test class. In fact it is detrimental, for the register could contain lons from parallel runs. Therefore the static field was made an instance field

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>