Prohibit usage of JUnit5 class lifecycle
JUnit5 offers to use the same class instance over several methods.
This however is not in line with the scenario concept of JGiven. The JGivenExtension removes the scenario after the exection of each state, resulting in failure if the next scenario is executed on the same instance

Moreover, it is unclear how scenarios that share the same instance react when these methods are executed in parallel.

In addition, the JUnit 5 testframework tests were adapted. It turns out that when JUnit5 fails on the class level it returns a successful report model for method but a failure for the class level.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>