Issue-829: Secure scenario case against concurrent modification
In general the Machinery of JGiven can be expected to be present in a single thread only. This is however not true for the Report model, which gets injected into all scenarios of a class. If a test with multiple test cases runs in parallel, then a single scenario model of a thread is taken as "lead" to aggregate all scenario cases. However, the same scenario model is still used to validate results of the test that run in the current thread.

This can not only lead to a "ConcurrentModificationException" when another thread tries to write its case into the model of the "lead thread" while said thread validates the results of its thread, but it could also lead to more tests being displayed as failed, than actually were.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>