Remove explicit support for Powermock
With the Upgrade of Mockito to Version 4.1, the current version of Powermock became incompatible with Mockito. A issue was filed in the powermock repo: https://github.com/powermock/powermock/issues/1112, but there was no reaction to it. In generar the repo appears in a derelict state an not even the question wether the repo is still alive could prompt a response from the maintainers. As compatibility is unlikely to be restored we remove the powermock test an thus the explicit powermock support. Existing integrations are likely to continue to work as long as users don't update their mockito.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>