Issue-829: Write Parallel JUnit 4 tests
The package with the dodgy name is actually a recommendation from the JUnit 4 site itself, therefore
I opted to use it here.

Contrary to JUnit5 I opted against having a second method in the classes, because paralellization works on class level only anyway which means that the method just lost what little use they had.

Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>