Issue-829: Throw an error if the usage of multithreaded tests with shared stages is detected in TestNG.
Also add documentation explaining why this is not supported. The documentation does not refer to testNG explicitly, because JUnit5 also offers a lifecyle model where the instance gets reused for multiple test methods. Likewise, the execution is prohibited there.

The point of execution of the incompatibility checker in TestNG was chosen after the JGiven test threw an error. Since other users could try to access the data in a similar fashion, I opted to not place the check at the start and instead make sure the data that is stored in the attributes is present (albeit useless).

Furthermore the check for multithreading on the method annotations has been discarded, because TestNG ignores that one anyway
Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>