Fix version not updated in kotlin plugin
by making the plugin declaration depend on the version specified in the properties file. However, this will make the example tests fail, because we do not and cannot publish a special plugin version for these. Therefore we need to hardcode the published version into the gradle file before setting the version to the test version.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>