Checkout actions fetches whole repository
Previously the repository did not checkout a single branch but actually a single commit, which it made into a unique ref for its own repo. Consequently, when attempting to rebase the entire history of gh-pages onto that commit the rebase would fail, because the base in the repo wasn't the actual master branch but only its latest commit which resulted in an unfitting history of that commit and a failing rebase.

Now that we check out everything from the start, this should work fine.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>