Fix Averaging in ProxyClassPerformanceTest
The reducer that we used here uses the first data point to determine the current average. Thus, when the there is an increment between the first and the second data point the increment will not get divided by the size of increments. This is incorrect and therefore we now divide all data points before summing up.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>