Make the maven plugin generation task dependent on the publishing task of its dependents
To produce the maven plugin we start a separate maven run. This maven run tries to find jgiven-core and jgiven-html-report as dependencies. However, since gradle does not know that this task depends on them being in the repository it may schedule the maven-generation task before the respective publishing tasks. This is especially true when the test task is active, as then the execution of jgiven-core takes way longer then the jgiven-maven-plugin.

I don't know exactly why this problem occurrs at this point. Maybe it was a fix in gradle that was backported...

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>