Prevent generateMavenTask from running all the time
In the previous configuration two publish tasks were declared as dependency of the generateMavenPlugin task which is itself a dependency of the jar task. Thus, when the conditional check for the generateMavenPlugin looks at the task graph it will always find a publish task in there, because configuring the generateMavenPlugin task put them there.
That obviously makes no sense.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>