Issue-1013: only write a version when there is a version to write
The error happens here, because we write the "compileOnly" dependencies manually.
In case of "compileOnly" dependencies gradle does not resolve any version. So for the JUnit5 dependencies gradle writes an empty version tag, producing an invalid pom.

However, gradle does correctly add the junit-bom to the pom as dependency management so that there should not be any version in that dependency in the first place.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>