Issue-802: Ensure commit is on graph
If the bot does not have admin rights on github, the push to master will fail. However, the creation of the tag will be successful creating a tag object that is not on the graph.
To avoid this we place the commit on a branch an push that branch, to be merged manually after the release

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>