Assert behavior of complex internal java types
I opted to fall back to the default behavior of inaccessible objects for Java types. The reason is that in contrast to primitive wrappers, introsepciton of a type like String would yield acutallly different information. Handing back the String (or BigInt) is a behavior that is not forseeable for the user, because the introspection result will not align with the actual value.

Signed-off-by: l-1sqared <30831153+l-1squared@users.noreply.github.com>