Fix task dependency conundrum in maven plugin
Unfortunately, the copy tasks copied into the main build directory, making all sorts of other tasks dependent on it in a nontrivial fashion. (I actually don't understand it, copy task should register the target as output; perhaps it clashes with targets that also use the folder as output?)
The simplest solution is to not make these two operations tasks that can be arranged freely.

Signed-off-by: l-1squared <30831153+l-1squared@users.noreply.github.com>